spin runAgent
