+++ ui.js
@@ -1,7 +1,6 @@
 // This file is the UI for the user. It accepts a TASK from the user and uses AI to complete the task. Tasks are related with code.
 const chalk = require('chalk');
 const agents = require('./agents');
-const prompts = require('prompts');
 const fs = require('fs');
 const path = require('path');
 
@@ -12,38 +11,9 @@ const { printGitDiff } = require('./modules/gitHelper');
 const { getFiles } = require('./modules/fsInput');
 const { generateAndWriteFileSummary } = require('./modules/summaries');
 const { getOptions } = require('./modules/cliOptions');
+const { runAgent } = require('./modules/interactiveAgent');
 
 const testingDirectory = '/benchmarks';
-
-/**
-@description Asynchronous function that runs an agent function with given variables.
-@param {function} agentFunction - The agent function to be executed asynchronously.
-@param {any} var1 - The first variable to be passed as an argument to the agent function.
-@param {any} var2 - The second variable to be passed as an argument to the agent function.
-@param {boolean} interactive=false - A boolean indicating whether or not to prompt the user for approval after running the agent function.
-@returns {Promise<any>} A Promise that resolves with the return value of the agent function if not in interactive mode, otherwise resolves or rejects based on user input.
-*/
-async function runAgent(agentFunction, var1, var2, interactive=false){
-  console.log("(agent)", agentFunction.name);
-  if (interactive){
-    res = await agentFunction(var1, var2);
-    console.dir(res, { depth: null })
-    const proceed = await prompts({
-      type: 'select',
-      name: 'value',
-      message: 'Approve agent\'s reply ?',
-      choices: [
-        { title: 'Approve - continue', value: 'continue' },
-        { title: 'Retry - Rerun agent', value: 'retry'},
-        { title: 'Abort', value: 'abort'}
-      ]
-    });
-    if (proceed.value === 'continue') return res
-    if (proceed.value === 'retry') await runAgent(agentFunction, var1, var2, interactive)
-    if (proceed.value === 'abort') throw new Error("Aborted")
-  }
-  return await agentFunction(var1, var2);
-}
 
 /**
  * Asynchronously reindexes the codebase located at the specified directory, using the specified model for indexing.
